Playlist Generation Outline
Created the basic outline of how Groupify will go about generating the big Groupify playlist off of its users. Also renamed the User class to GroupifyUser to avoid naming collision with the Spotify Web API and changed the top tracks and top artist fields to be Lists instead of arrays for easier indexing and searching.